move motor commands to 'command' folder for consistency
